requirements/requirements.txt
datasets
Pillow
numpy
PyYAML
ultralytics